Refactored animalQuantitiesDiff to helpers and cleaned up unused code
